# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.TotalCostResponse:
    properties:
      total:
        type: integer
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST-сервис для управления подписками пользователей
  title: Subscriptions Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список всех подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Поля для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      parameters:
      - description: Начало периода (MM-YYYY)
        in: query
        name: from
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: to
        type: string
      - description: Фильтр по user_id
        in: query
        name: user_id
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TotalCostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Суммарная стоимость подписок за период
      tags:
      - subscriptions
swagger: "2.0"